+++ packages/plugin-utils/src/__test__/Identifier.test.ts
@@ -1,0 +1,9 @@
+import { CONFIG_KEYS } from '../Identifier';
+
+const pluginName = 'utils';
+test('CONFIG_KEYS CFG_KEY', () => {
+  expect(CONFIG_KEYS.CFG_KEY(pluginName)).toBe(`CONFIG:KEY:${pluginName}`);
+});
+test('CONFIG_KEYS PLUGIN_CFG_KEY', () => {
+  expect(CONFIG_KEYS.PLUGIN_CFG_KEY(pluginName)).toBe(`CONFIG:PLUGIN:${pluginName}`);
+});