+++ packages/cli/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.4](https://github.com/Jetsly/ddot/compare/@ddot/ddot-cli@0.0.3...@ddot/ddot-cli@0.0.4) (2018-11-30)
+
+
+### Features
+
+* **ddot-plugin-webpack:** export document.title ([0c4748c](https://github.com/Jetsly/ddot/commit/0c4748c))
+
+
+
+
+
+## [0.0.3](https://github.com/Jetsly/ddot/compare/@ddot/ddot-cli@0.0.2...@ddot/ddot-cli@0.0.3) (2018-11-29)
+
+**Note:** Version bump only for package @ddot/ddot-cli
+
+
+
+
+
 ## [0.0.2](https://github.com/Jetsly/ddot/compare/@ddot/ddot-cli@0.0.1...@ddot/ddot-cli@0.0.2) (2018-11-21)